requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
roboflow>=1.1.0
streamlit>=1.40.0
Pillow>=10.0.0
numpy>=1.26.0
PyYAML>=6.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
python-dotenv>=1.0.0
